Project Restructuring
- Moved pieces and related classes to jchess.pieces
- Moved UI-related files, excluding JChessView and GUI, to jchess.UI
- Moved classes Chessboard and Square to jchess.UI.board
