+++ config.txt
@@ -1,2 +1,2 @@
-#Mon Oct 28 20:35:16 CET 2019
-THEME=matlak
+#Thu Nov 07 22:45:46 CET 2019
+THEME=hunter
