+++ jchess/JChessApp.java
@@ -22,7 +22,7 @@ import org.jdesktop.application.SingleFrameApplication;
  * The main class of the application.
  */
 public class JChessApp extends SingleFrameApplication {
-     static JChessView jcv;
+     public static JChessView jcv;
     /**
      * At startup create and show the main frame of the application.
      */
