+++ jchess/JChessView.java
@@ -20,6 +20,13 @@ import org.jdesktop.application.ResourceMap;
 import org.jdesktop.application.SingleFrameApplication;
 import org.jdesktop.application.FrameView;
 import org.jdesktop.application.TaskMonitor;
+
+import jchess.UI.JChessAboutBox;
+import jchess.UI.JChessTabbedPane;
+import jchess.UI.NewGameWindow;
+import jchess.UI.PawnPromotionWindow;
+import jchess.UI.ThemeChooseWindow;
+
 import java.awt.event.ActionEvent;
 import java.awt.event.ActionListener;
 import javax.swing.Timer;
@@ -241,7 +248,7 @@ public class JChessView extends FrameView implements ActionListener, ComponentLi
     private void initComponents() {
 
         mainPanel = new javax.swing.JPanel();
-        gamesPane = new jchess.JChessTabbedPane();
+        gamesPane = new jchess.UI.JChessTabbedPane();
         menuBar = new javax.swing.JMenuBar();
         javax.swing.JMenu fileMenu = new javax.swing.JMenu();
         newGameItem = new javax.swing.JMenuItem();
@@ -576,7 +583,7 @@ public class JChessView extends FrameView implements ActionListener, ComponentLi
         throw new UnsupportedOperationException("Not supported yet.");
     }
     
-    protected Game getActiveTabGame() throws ArrayIndexOutOfBoundsException
+    public Game getActiveTabGame() throws ArrayIndexOutOfBoundsException
     {
         Game activeGame = (Game)this.gamesPane.getComponentAt(this.gamesPane.getSelectedIndex());
         return activeGame;
