+++ jchess/Player.java
@@ -29,14 +29,14 @@ import java.io.Serializable;
 public class Player implements Serializable
 {
 
-    String name;
+    public String name;
 
-    enum colors
+    public enum colors
     {
 
         white, black
     }
-    colors color;
+    public colors color;
 
     public enum playerTypes
     {
@@ -44,7 +44,7 @@ public class Player implements Serializable
         localUser, networkUser, computer
     }
     public playerTypes playerType;
-    boolean goDown;
+    public boolean goDown;
 
     public Player()
     {
@@ -68,7 +68,7 @@ public class Player implements Serializable
     /** Method getting the players name
      *  @return name of player
      */
-    String getName()
+    public String getName()
     {
         return this.name;
     }
