+++ jchess/UI/Clock.java
@@ -18,7 +18,9 @@
  * Mateusz Sławomir Lach ( matlak, msl )
  * Damian Marciniak
  */
-package jchess;
+package jchess.UI;
+
+import jchess.Player;
 
 /** Class to represent seperate wall-clock for one player.
  *  Full ChessClock is represented by GameClock object (two clock - one for each player)
