+++ jchess/UI/DrawLocalSettings.java
@@ -17,7 +17,7 @@
  * Authors:
  * Mateusz Sławomir Lach ( matlak, msl )
  */
-package jchess;
+package jchess.UI;
 
 import javax.swing.*;
 import java.awt.event.ActionListener;
@@ -26,6 +26,14 @@ import java.awt.event.TextListener;
 import java.awt.event.TextEvent;
 import java.awt.*;
 import javax.swing.text.BadLocationException;
+
+import jchess.Game;
+import jchess.JChessApp;
+import jchess.Player;
+import jchess.Settings;
+import jchess.Player.playerTypes;
+import jchess.Settings.gameModes;
+import jchess.Settings.gameTypes;
 
 /**
  * Class responsible for drawing the fold with local game settings
@@ -189,7 +197,7 @@ public class DrawLocalSettings extends JPanel implements ActionListener, TextLis
 
     }
 
-    DrawLocalSettings(JDialog parent)
+    public DrawLocalSettings(JDialog parent)
     {
         super();
         //this.setA//choose oponent
