+++ jchess/UI/board/Square.java
@@ -18,7 +18,9 @@
  * Mateusz Sławomir Lach ( matlak, msl )
  * Damian Marciniak
  */
-package jchess;
+package jchess.UI.board;
+
+import jchess.pieces.Piece;
 
 /**
  * Class to represent a chessboard square
@@ -26,8 +28,8 @@ package jchess;
 public class Square
 {
 
-    int pozX; // 0-7, becouse 8 squares for row/column
-    int pozY; // 0-7, becouse 8 squares for row/column
+    public int pozX; // 0-7, becouse 8 squares for row/column
+    public int pozY; // 0-7, becouse 8 squares for row/column
     public Piece piece = null;//object Piece on square (and extending Piecie)
 
     Square(int pozX, int pozY, Piece piece)
@@ -38,7 +40,7 @@ public class Square
     }/*--endOf-Square--*/
 
 
-    Square(Square square)
+    public Square(Square square)
     {
         this.pozX = square.pozX;
         this.pozY = square.pozY;
