+++ jchess/pieces/Knight.java
@@ -18,9 +18,15 @@
  * Mateusz Sławomir Lach ( matlak, msl )
  * Damian Marciniak
  */
-package jchess;
+package jchess.pieces;
 
 import java.util.ArrayList;
+
+import jchess.GUI;
+import jchess.Player;
+import jchess.Player.colors;
+import jchess.UI.board.Chessboard;
+
 import java.awt.Graphics;
 import java.awt.Image;
 
@@ -34,7 +40,7 @@ public class Knight extends Piece
     protected static final Image imageWhite = GUI.loadImage("Knight-W.png");
     protected static final Image imageBlack = GUI.loadImage("Knight-B.png");
 
-    Knight(Chessboard chessboard, Player player)
+    public Knight(Chessboard chessboard, Player player)
     {
         super(chessboard, player);//call initializer of super type: Piece
         //this.setImages("Knight-W.png", "Knight-B.png");
