+++ jchess/pieces/Rook.java
@@ -19,9 +19,15 @@
  * Damian Marciniak
  */
 
-package jchess;
+package jchess.pieces;
 
 import java.util.ArrayList;
+
+import jchess.GUI;
+import jchess.Player;
+import jchess.Player.colors;
+import jchess.UI.board.Chessboard;
+
 import java.awt.Graphics;
 import java.awt.Image;
 
@@ -43,12 +49,12 @@ import java.awt.Image;
 public class Rook extends Piece
 {
 
-    boolean wasMotion = false;
+    public boolean wasMotion = false;
     protected static final Image imageWhite = GUI.loadImage("Rook-W.png");
     protected static final Image imageBlack = GUI.loadImage("Rook-B.png");
     public static short value = 5;
 
-    Rook(Chessboard chessboard, Player player)
+    public Rook(Chessboard chessboard, Player player)
     {
         super(chessboard, player);//call initializer of super type: Piece
         //this.setImages("Rook-W.png", "Rook-B.png");
