+++ jchess/server/Console.java
@@ -24,7 +24,6 @@ import java.io.IOException;
 import java.util.Map;
 import java.util.logging.Level;
 import java.util.logging.Logger;
-import jchess.MD5;
 
 public class Console
 {
