+++ jchess/server/Server.java
@@ -113,9 +113,9 @@ public class Server implements Runnable
                 }
                 Table table = tables.get(tableID);
 
-                if (!jchess.MD5.encrypt(table.password).equals(password))
+                if (!jchess.server.MD5.encrypt(table.password).equals(password))
                 {
-                    print("bad password: " + jchess.MD5.encrypt(table.password) + " != " + password);
+                    print("bad password: " + jchess.server.MD5.encrypt(table.password) + " != " + password);
                     output.writeInt(Connection_info.err_bad_password.getValue());
                     output.flush();
                     continue;
